# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cotonou Municipal Garage API",
        "description": "API documentation for the Cotonou Municipal Garage System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/procedures": {
            "get": {
                "tags": [
                    "Procedures"
                ],
                "summary": "Get list of procedures",
                "description": "Returns list of procedures",
                "operationId": "getProceduresList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "procedure_type",
                        "in": "query",
                        "description": "Filter by procedure type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter by end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProcedureResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Procedures"
                ],
                "summary": "Store new procedure",
                "description": "Returns procedure data",
                "operationId": "storeProcedure",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProcedureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcedureResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/procedures/{id}": {
            "get": {
                "tags": [
                    "Procedures"
                ],
                "summary": "Get procedure information",
                "description": "Returns procedure data",
                "operationId": "getProcedureById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of procedure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcedureResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Procedures"
                ],
                "summary": "Update existing procedure",
                "description": "Returns updated procedure data",
                "operationId": "updateProcedure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of procedure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProcedureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcedureResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Procedures"
                ],
                "summary": "Delete existing procedure",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProcedure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of procedure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProcedureRequest": {
                "title": "Store Procedure Request",
                "required": [
                    "vehicle_id",
                    "procedure_type",
                    "cost"
                ],
                "properties": {
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_type": {
                        "type": "string",
                        "enum": [
                            "impound",
                            "release",
                            "auction"
                        ],
                        "example": "impound"
                    },
                    "description": {
                        "type": "string",
                        "example": "Vehicle impounded for illegal parking.",
                        "nullable": true
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    }
                },
                "type": "object"
            },
            "UpdateProcedureRequest": {
                "title": "Update Procedure Request",
                "properties": {
                    "procedure_type": {
                        "type": "string",
                        "enum": [
                            "impound",
                            "release",
                            "auction"
                        ],
                        "example": "release"
                    },
                    "description": {
                        "type": "string",
                        "example": "Vehicle released after payment.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "example": "completed"
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 80
                    }
                },
                "type": "object"
            },
...
            "ProcedureDocumentResource": {
                "title": "Procedure Document Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "document_type": {
                        "type": "string",
                        "example": "impound_report"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "documents/impound_report_1.pdf"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/documents/impound_report_1.pdf"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProcedureResource": {
                "title": "Procedure Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_type": {
                        "type": "string",
                        "enum": [
                            "impound",
                            "release",
                            "auction"
                        ],
                        "example": "impound"
                    },
                    "description": {
                        "type": "string",
                        "example": "Vehicle impounded for illegal parking."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "documents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProcedureDocumentResource"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "ulid",
                        "example": "01H8XGJWBWBAQ8PEJ9C3J745A1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "manager",
                            "agent"
                        ],
                        "example": "agent"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}